Add MetaFields.java
